update a little bit: use float instead of int. The goal of the test is not changed. But hard-coded code to get buffer element size based on the illusion of only integer type would break on the test.

+++ c/gep_const_offset_1D.c
@@ -15,7 +15,7 @@ main(int argc, char **argv) {
   }
   {
     unsigned tmp;
-    unsigned buffer[4] = { 0, 0, 0, 0};
+    float buffer[4] = { 0, 0, 0, 0};
     buffer[2] = 1;
     tmp = buffer[2];
     buffer[-1] = 2;
